Bookedrooms.php getting updated with bookings. Booking done.

+++ admin/rooms.php
@@ -14,8 +14,8 @@ if (!$dbc) {
    // code...
     $number = $_POST['number'];
     $type = $_POST['type'];
-    $position= $_POST['position'];
-    $query = "INSERT INTO rooms (room,type,roomposition) VALUES ('$number','$type','$position')";
+    $floor= $_POST['floor'];
+    $query = "INSERT INTO rooms (room,type,floor) VALUES ('$number','$type','$floor')";
     $update_status = mysqli_query($dbc, $query);
     // var_dump($update_status);
     if(!$update_status){
@@ -28,7 +28,7 @@ if (!$dbc) {
           'Successfully Updated.<button type="button" class="close" data-dismiss="alert" aria-label="Close">' .
           '<span aria-hidden="true">&times;</span></button></div></div>';
     }
-    $activeTab = "2";
+    $activeTab = "3";
 }
 
 
@@ -55,7 +55,7 @@ if (!$dbc) {
           'Successfully Updated.<button type="button" class="close" data-dismiss="alert" aria-label="Close">' .
           '<span aria-hidden="true">&times;</span></button></div></div>';
     }
-    $activeTab = "2";
+    $activeTab = "3";
   }
  ?>
 
@@ -70,7 +70,10 @@ if (!$dbc) {
        <a class="nav-link <?php if($activeTab==2){echo 'active';} ?>" id="profile-tab" data-toggle="tab" href="#empty" role="tab" aria-controls="profile" aria-selected="false">Empty</a>
      </li>
      <li class="nav-item pill-1">
-       <a class="nav-link <?php if($activeTab==3){echo 'active';} ?>" id="home-tab" data-toggle="tab" href="#add" role="tab" aria-controls="home" aria-selected="true">Add Rooms </a>
+       <a class="nav-link <?php if($activeTab==3){echo 'active';} ?>" id="home-tab" data-toggle="tab" href="#all" role="tab" aria-controls="home" aria-selected="true">All Rooms </a>
+     </li>
+     <li class="nav-item pill-1">
+       <a class="nav-link <?php if($activeTab==4){echo 'active';} ?>" id="home-tab" data-toggle="tab" href="#add" role="tab" aria-controls="home" aria-selected="true">Add Rooms </a>
      </li>
    </ul>
    <div class="tab-content" id="companiesTabContent">
@@ -82,10 +85,9 @@ if (!$dbc) {
              <th scope="col">S.No.</th>
              <th scope="col">Room Number</th>
              <th scope="col">Room Type</th>
-             <th scope="col">Status</th>
              <th scope="col">Id</th>
              <th scope="col">Guest Name</th>
-             <th scope="col">Position</th>
+             <th scope="col">Room Floor Number</th>
              <th scope="col">Username</th>
              <th scope="col">Indentor Name</th>
              <th scope="col">Check-In Date</th>
@@ -99,7 +101,7 @@ if (!$dbc) {
            die("Connection failed: " . mysqli_connect_error());
          }
 
-           $query = "SELECT room,type,status,id,guestname,roomposition,username,indentorname,arrival,departure FROM rooms WHERE status='booked'";
+           $query = "SELECT room,type,id,guestname,floor,username,indentorname,arrival,departure FROM rooms WHERE status='booked'";
            $data = mysqli_query($dbc, $query);
            if(mysqli_num_rows($data) != 0){
          ?>
@@ -110,10 +112,9 @@ if (!$dbc) {
                echo '<tr><th scope="row">' . $curr . '</th>' .
                          '<td>' . $row["room"] . '</td>' .
                          '<td>' . $row["type"] . '</td>' .
-                         '<td>' . $row["status"] . '</td>' .
                          '<td>' . $row["id"] . '</td>' .
                          '<td>' . $row["guestname"] . '</td>' .
-                         '<td>' . $row["roomposition"] . '</td>' .
+                         '<td>' . $row["floor"] . '</td>' .
                          '<td>' . $row["username"] . '</td>' .
                          '<td>' . $row["indentorname"] . '</td>' .
                          '<td>' . $row["arrival"] . '</td>' .
@@ -138,9 +139,7 @@ if (!$dbc) {
              <th scope="col">S.No.</th>
              <th scope="col">Room Number</th>
              <th scope="col">Room Type</th>
-             <th scope="col">Status</th>
-             <th scope="col">Position</th>
-             <th scope="col">Delete</th>
+             <th scope="col">Room Floor Number</th>
            </tr>
          </thead>
          <?php
@@ -148,7 +147,7 @@ if (!$dbc) {
          if (!$dbc) {
            die("Connection failed: " . mysqli_connect_error());
          }
-             $query = "SELECT serial,room,type,status,id,guestname,roomposition,username,indentorname,arrival,departure FROM rooms WHERE status='empty'";
+             $query = "SELECT serial,room,type,status,id,guestname,floor,username,indentorname,arrival,departure FROM rooms WHERE status='empty'";
            $data = mysqli_query($dbc, $query);
            if(mysqli_num_rows($data) != 0){
          ?>
@@ -159,9 +158,50 @@ if (!$dbc) {
                echo '<tr><th scope="row">' . $curr . '</th>' .
                          '<td>' . $row["room"] . '</td>' .
                          '<td>' . $row["type"] . '</td>' .
-                         '<td>' . $row["status"] . '</td>' .
-                         '<td>' . $row["roomposition"] . '</td>' .
-                         '<td><form action="' . $_SERVER['PHP_SELF'] . '?serial=' . $row["serial"] . '&tab=3" method="post">' .
+                         '<td>' . $row["floor"] . '</td>' .
+                     '</tr>';
+               $curr = $curr + 1;
+             }
+           ?>
+         </tbody>
+         <?php } else { ?>
+           <tr>
+             <td>No data</td>
+           </tr>
+         <?php } ?>
+       </table>
+
+     </div>
+
+     <div class="tab-pane fade <?php if($activeTab==3){echo 'show active';} ?>" id="all" role="tabpanel">
+       <table class="table">
+         <thead class="thead-light">
+           <tr>
+             <th scope="col">S.No.</th>
+             <th scope="col">Room Number</th>
+             <th scope="col">Room Type</th>
+             <th scope="col">Room Floor Number</th>
+             <th scope="col">Delete</th>
+           </tr>
+         </thead>
+         <?php
+         $dbc= mysqli_connect('localhost','root','','guesthouse');
+         if (!$dbc) {
+           die("Connection failed: " . mysqli_connect_error());
+         }
+             $query = "SELECT room,type,floor FROM rooms";
+           $data = mysqli_query($dbc, $query);
+           if(mysqli_num_rows($data) != 0){
+         ?>
+         <tbody>
+           <?php
+             $curr = 1;
+             while($row = mysqli_fetch_array($data)){
+               echo '<tr><th scope="row">' . $curr . '</th>' .
+                         '<td>' . $row["room"] . '</td>' .
+                         '<td>' . $row["type"] . '</td>' .
+                         '<td>' . $row["floor"] . '</td>' .
+                         '<td><form action="' . $_SERVER['PHP_SELF'] . '?room=' . $row["room"] . '&tab=3" method="post">' .
                          '<button type="delete" class="btn btn-outline-danger" name="delete">Delete</button></form></td>' .
                      '</tr>';
                $curr = $curr + 1;
@@ -177,13 +217,13 @@ if (!$dbc) {
 
      </div>
 
-     <div class="tab-pane fade <?php if($activeTab==3){echo 'show active';} ?>" id="add" role="tabpanel">
+     <div class="tab-pane fade <?php if($activeTab==4){echo 'show active';} ?>" id="add" role="tabpanel">
        <table class="table">
          <thead class="thead-light">
            <tr>
              <th scope="col">Room Number</th>
              <th scope="col">Room Type</th>
-             <th scope="col">Room Position</th>
+             <th scope="col">Room Floor Number</th>
              <th scope="col">Submit</th>
            </tr>
          </thead>
@@ -203,7 +243,7 @@ if (!$dbc) {
                  </td>
                  <td>
                  <div class="col">
-                   <input type="text" class="form-control" placeholder="Room Position"  name="position" required>
+                   <input type="text" class="form-control" placeholder="Room floor"  name="floor" required>
                  </div>
                  </td>
                  <td>
